### 2024-06-15 #6893 @editorialbot
[PRE REVIEW]: MACE: a Machine-learning Approach to Chemistry Emulation
**Submitting author:** @silkemaes (<a href="http://orcid.org/0000-0003-4159-9964">Silke Maes</a>)
**Repository:** https://github.com/silkemaes/MACE
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.0.0
