### 2024-07-29 #6893 @silkemaes
Hi @JBorrow! Sorry for my late reply, I was on holidays until last week.

You are right, that information is outdated. I will update it. The submission to JOSS indeed for me was like a release for the final version. Though, I didn't officially release this version yet, since I was planning to do this after the JOSS reviewing process. Is that OK, or would it be better that I make a release of this version on GitHub?
