### 2024-07-29 #6893 @JBorrow
I would suggest making a release that our reviewers can look at (i.e. they can look at the pinned version you created). That way, you can make a new release on publication that will show the 'before and after' JOSS reviews.
