### 2024-07-30 #6893 @silkemaes
I just made a release of the branch JOSS that contains the version of MACE submitted to JOSS.
