### 2024-08-06 60e8141
add requirements.txt based on documentation
