### 2024-08-06 95c2b3e
Add scipy as a requirement, since MACE doesn't run without it.
