+++ .github/workflows/lint-black.yml
@@ -1,13 +1,0 @@
-name: Lint_with_black
-
-on: [push, pull_request]
-
-jobs:
-  lint:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-      - uses: psf/black@stable
-        with:
-          options: "--check --verbose"
-          src: "./src"
