### 2024-08-16 64b596d
Checking robustness + fixing minor bugs
