### 2024-08-26 #6893 @JBorrow
Hi @trappitsch; I found your name in the JOSS reviewer database. Would you be interested in reviewing this submission for us? You can see our review guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
