### 2024-08-27 #7148 @editorialbot
[REVIEW]: MACE: a Machine-learning Approach to Chemistry Emulation
**Submitting author:** @silkemaes (<a href="http://orcid.org/0000-0003-4159-9964">Silke Maes</a>)
**Repository:** https://github.com/silkemaes/MACE
**Branch with paper.md** (empty if default branch): JOSS
**Version:** v1.1.0
