### 2024-09-11 f07e3da
Adjusting path to training data
