+++ src/mace/CSE_0D/dataset.py
@@ -312,7 +312,7 @@ class CSEmod():
         '''
 
         if not inpackage:
-            self.path = '/STER/silkem/CSEchem/' + path[34:-17]
+            self.path = '/STER/silkem/CHEM/out/' + path[34:-17]
             self.model = path[34:-51]
             self.name = path[-43:-18]
             inp_path = self.path[:-26]+ 'inputChemistry_'+self.name+'.txt'
