### 2024-09-19 silkemaes/MACE#7 @richings
JOSS Review: minor issues with software installation and documentation
Hi @silkemaes. I am reviewing your MACE software paper for JOSS (the JOSS review thread can be found [here](https://github.com/openjournals/joss-reviews/issues/7148)

I am trying to set up MACE so I can run the software and test the reproducibility of the results. However, I had a few questions about the Documentation and how to install the software:

- Is there an automated way to install all of the dependencies/pre-requisites? I was finding I had to pip install all the pre-requisites (e.g. torch) manually, but I wasn't sure if there was a list somewhere with all of the dependencies (sorry if I've missed this!). I think I saw that Steven Rieder has an open pull request to add a "requirements.txt" file, which I think should solve this.

- In the README you have a few paragraphs on How to Run, but I wasn't sure exactly how to do this. For example, it says about setting up the Dataset class, and it points to the dataset.py file, but I wasn't sure how to set up the Dataset class. I tried to just run the run.py script, but it gives a File Not Found error as it couldn't load the file MACE/data/rate16.specs. Do I need to download this file? Sorry, I couldn't find it on the repository.

- It looks like there is a Sphinx "Read the Docs" documentation (e.g. I can see a readthedocs.yaml file), but I wasn't sure where to find the documentation itself? Do I need to "compile/build" this documentation as it were? (Sorry if I've missed it!).

- I saw there were a few Jupter Notebooks in the repository as well, are these designed to be introductory tutorials for new users to help learn how to use the code? I tried running the robustness.ipynb notebook, but I encountered the same error with the rate16.specs file not found.
