### 2024-09-19 #7148 @richings
I had a few minor issues with the software installation and the documentation; I've opened an issue for this on the target repository [here](https://github.com/silkemaes/MACE/issues/7).
