### 2024-09-26 silkemaes/MACE#10 @trappitsch
JOSS Review: No test suite
Testing is an integral part of software and a requirements for JOSS submission, see [here](https://joss.readthedocs.io/en/latest/submitting.html#submitting-a-paper-to-joss). While I saw that `pytest` is listed as a requirements, it can't find any tests and neither did I looking manually. I might have missed it, could you clarify?

Details on the test suite that needs to pass should also go into the documentation.
