### 2024-10-01 f4f3d6d
Update pyproject.toml
