### 2024-10-10 980762a
Adding required packages via file & removing numba
