### 2024-10-10 3a610d8
Adjusting README and files according to JOSS reviewers
