+++ input/example.in
@@ -25,11 +25,3 @@ elm = 0
 
 
 
-
-Name = 20240604_133421
-
-Name = 20240604_133421
-
-Name = 20240604_133538
-
-Name = 20240604_133932
