### 2024-10-10 f2518c1
Adjusting README and files according to JOSS reviewers
