+++ docs/src/conf.py
@@ -37,7 +37,9 @@ extensions = [
     'sphinx.ext.autosummary',
     'sphinx.ext.napoleon',
     'sphinx_copybutton',
+    'sphinxcontrib.email',
     'nbsphinx'
+    
 ]
 
 # Add any paths that contain templates here, relative to this directory.
