### 2024-10-10 silkemaes/MACE#7 @silkemaes
Hi @richings.

First of all thank you for reviewing, and sorry for the late response. First of all, I don't know if it was clear, but the aim was for the JOSS reviewers to review the code in the branch JOSS.

- There is no automated way to install MACE currently, it is best to just clone the repo. I added this in the readme file, and merged the pull request of Steven Rieder for the ```requirements.txt```.
- You are right that is was not very clear how to run MACE. I now adjusted the explanation in the readme. I renamed the ```run.py``` file to ```routine.py``` in order to mitigate some confusion. Namely, MACE cannot just be 'ran', it can only be trained. The routine to train MACE is the current ```routine.py``` file. About the ```Dataset``` class, when training an architecture with PyTorch, the data needs to be in the correct tensor form. The script ```dataset.py``` makes sure that my data is converted in the correct way for PyTorch to use it for training. However, this script is only applicable for my specific case. If someone wants to train a MACE architecture themselves, they first need to rewrite the ```dataset.py``` script. Hence, It is better to test a trained MACE model using the notebook in the [Documentation](https://mace-code.readthedocs.io/en/latest/use/index.html).
- The documentation is available online: https://mace-code.readthedocs.io/.
- The notebook on robustness does not need to be reviewed, since this was not originally submitted to JOSS.
