+++ docs/src/conf.py
@@ -1,73 +1,0 @@
-
-# Configuration file for the Sphinx documentation builder.
-#
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
-
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-
-import os
-import sys
-sys.path.insert(0, os.path.abspath('.'))
-import datetime
-year = datetime.datetime.now().year
-
-# -- Project information -----------------------------------------------------
-
-project = 'MACE'
-copyright = f'{year}, Silke Maes'
-author = 'Silke Maes'
-
-# -- Explicitly specify master doc -------------------------------------------
-master_doc = 'index'
-
-
-# -- General configuration ---------------------------------------------------
-
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
-extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.autosummary',
-    'sphinx.ext.napoleon',
-    'sphinx_copybutton',
-    'nbsphinx'
-]
-
-# Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
-
-# List of patterns, relative to source directory, that match files and
-# directories to ignore when looking for source files.
-# This pattern also affects html_static_path and html_extra_path.
-exclude_patterns = ['_build']
-
-
-# -- Options for HTML output -------------------------------------------------
-
-# The theme to use for HTML and HTML Help pages.  See the documentation for
-# a list of builtin themes.
-#
-html_theme = "piccolo_theme"
-# html_logo  = 'images/pomme_logo.svg'
-# html_theme_options = {'logo_only': True}
-# html_sidebars = {'**': ['util/searchbox.html', 'util/sidetoc.html']}
-
-
-# Add any paths that contain custom static files (such as style sheets) here,
-# relative to this directory. They are copied after the builtin static files,
-# so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static', 'images']
-
-
-# -- Sphinx Python API docs --------------------------------------------------
-
-autodoc_member_order = 'groupwise'
-autodoc_default_flags = ['members']
-autosummary_generate = True
