+++ docs/src/example/index.rst
@@ -1,13 +1,0 @@
-
-
-Examples
-#####################
-
-On this page shows tutorials on how to train a MACE model and how to test a trained model.
-
-.. toctree::
-   :maxdepth: 1
-
-   run.ipynb
-   load&test.ipynb
-
