### 2024-10-10 silkemaes/MACE#10 @silkemaes
It is true that we haven't written automated tests for MACE. I know JOSS strongly encourages to write tests, however since I am not experienced with automated testing, currently we have not provided these tests.

Do you want me to provide some tests? Then I will look into it.
