### 2024-10-18 silkemaes/MACE#7 @richings
Hi @silkemaes

Many thanks for your response! Sorry, you're right, I was looking at the wrong branch! I'll switch to the JOSS branch for the review.

The updates to the README are much clearer, thanks for the clarifications! There was just one minor thing - in the "How to use?" section you include a link to an annotated notebook to show how to use the routine for training the MACE model. However, I think this link is broken, it gives me a 404 error? I was able to find the notebook on the mace-code.readthedocs.io web page though, I think it's just an incorrect url in the README.

The link to the other annotated notebook, in the "Example case" section of the README, does work fine though.
