### 2024-10-18 silkemaes/MACE#14 @richings
JOSS Review: Issues with training the MACE model using the annotated notebook
Hi @silkemaes

I have tried to run the annotated notebook for training a MACE model that you provide in the Documentation here:

https://mace-code.readthedocs.io/en/latest/use/routine.html

(This is the one that's based on the `routine.py` script). However, I had a couple of problems with running this notebook that meant I was unable to train the model.

Firstly, when I import the dataset using:

`import src.mace.CSE_0D.dataset  as ds`

It gives me an error message that it can't find `MACE/data/rate16.specs`

I can see the `rate16.specs` file is in the parent `MACE/` directory, so I created a new sub-directory `MACE/data` and put it in there. Is this correct? That did at least avoid the file not found error, so it was able to run. Would it be worth moving this file into the "data" directory in the repository too, to make it easier to find when we run the notebook? Also, the `rate16.specs` file that I got from the repository (on the JOSS branch) is empty, is this correct?

I then got as far as loading the train, validation and test datasets, but when I try to load these using the `ds.get_data` function it gives me the following error:

`FileNotFoundError: /STER/silkem/MACE/data/paths_data_C.txt not found.`

There are a couple of issues here:

- It looks like it is searching for the file in the MACE directory located in `/STER/silkem/MACE`. I tried to update the paths in the Jupyter Notebook that you gave in the documentation so that the paths point to the correct directories on my machine, but I think there are some places in the code where it is still hard-coded to the path on your machine? E.g. in `src/mace/CSE_0D/dataset.py` on line 85 it has the following:

`loc = '/STER/silkem/MACE/'`

I think this is why it is looking for the file in the wrong directory?

- Secondly, it is looking for a file called `paths_data_C.txt`, but I'm not sure where this file is, I couldn't see it in the repository? Is there another step that I need to do beforehand to create or set up this file? Sorry if I missed it!
