### 2024-10-18 #7148 @richings
I had some problems with running the example annotated notebook for training the MACE model that's provided in the documentation. I've opened a new issue on the target repository [here](https://github.com/silkemaes/MACE/issues/14).
