+++ README.md
@@ -52,7 +52,7 @@ For more details, check out our paper: [Maes et al. (2024)](https://ui.adsabs.ha
 
 The script ```routine.py``` gives the flow of training & storing a MACE architecture, and immediately applies to the specified test dataset once training is finished. As such, it returns an averaged error on the MACE model compared to the classical model. More info on the training routine can be found in the [paper](https://ui.adsabs.harvard.edu/abs/2024arXiv240503274M/abstract). 
 
-An annotated notebook of the routine can be found in the [documentation](https://mace-code.readthedocs.io/en/latest/example/run.html).
+An annotated notebook of the routine can be found in the [documentation](https://mace-code.readthedocs.io/en/latest/use/index.html).
 
 The script ```routine.py``` takes an input file with the needed (hyper)parameter setup. An example of such an input file can be found in input/.
 ```
@@ -70,7 +70,7 @@ should be made compatible with that data. Currently, the script ```src/mace/CSE_
 
 This repository contains a trained MACE model as a test case, see ```model/20240604_160152```. 
 
-The code for loading a trained MACE model can be found in the script ```src/mace/load.py```, testing in ```src/mace/test.py```. An annotated notebook can be found in the [documentation](https://mace-code.readthedocs.io/en/latest/example/load%26test.html).
+The code for loading a trained MACE model can be found in the script ```src/mace/load.py```, testing in ```src/mace/test.py```. An annotated notebook can be found in the [documentation](https://mace-code.readthedocs.io/en/latest/example/index.html).
 
 ---
 ## Contact <a name="cont"></a>
