### 2024-10-21 silkemaes/MACE#7 @silkemaes
Hi @richings,

You are right! With changing a bit the documentation, the link changed. I fixed this now.
