+++ data/train/20211014_gridC_Mdot1e-6_v12-5_T_eps_model_2022-12-23h10-52-30/inputChemistry_model_2022-12-23h10-52-30.txt
@@ -1,0 +1,57 @@
+! INPUT PARAMETERS FOR THE RATE13_CSE_MODEL
+!
+! PHYSICAL PARAMETERS OF THE CIRCUMSTELLAR ENVELOPE
+R_STAR = 20000000000000
+T_STAR = 2050.0
+MLOSS = 1e-06
+VELOCITY = 1250000
+TEMPMODE = EPSILON
+EPSILON_TEMP_LAW = 0.3
+R_EPSIL = 20000000000000
+!
+! GRAIN PARAMETERS FOR H2 FORMATION AND CR IONISATION
+A_G = 1e-05
+X_G = 1.5e-12
+ALBEDO = 0.5
+ZETA = 1
+GAMCO = 3
+AUV_AV = 4.65
+!
+! RADIUS INFORMATION
+R_INNER_CHEM = 100000000000000
+R_OUTER_CHEM = 1000000000000000000
+RESOLUTION = 0.03
+!
+! ANALYSE SUBROUTINE
+PERFORM_ROUTINE = 0
+ROUTINE_RADIUS = 0
+FULL_ANALYSIS = 0
+SENSITIVITY = 0
+!
+! DVODE
+RTOL = 1e-05
+ATOL = 1e-25
+!
+! CLUMPING PARAMETERS
+CLUMPMODE = SMOOTH
+FVOL = 0
+L = 0
+FIC = 0
+!
+! INNER PHOTONS
+ISTELLAR = 0
+IBIN = 0
+RSCALE = 50
+RDUST = 2
+TBIN = 6000
+RBIN = 0
+!
+! SHIELDING FUNCTIONS CO AND N2
+FILECO = "/lhome/silkem/CHEM/src-IP-AP-HNR/shielding/co/shield.03.5.69-557-36.dat"
+FILEN2 = "/lhome/silkem/CHEM/src-IP-AP-HNR/n2/Tex_10K_bH2_3km.s-1_bH_5km.s-1_NH_1e+20cm-2.dat"
+! INPUT AND OUTPUT FILE NAMES
+REACTIONS_FILE = "/lhome/silkem/CHEM/src-IP-AP-HNR/rates/rate16_IP_2330K_AP_6000K_corrected.rates"
+SPECIES_FILE = "/lhome/silkem/CHEM/src-IP-AP-HNR/specs/rate16_IP_6000K_Crich_mean_Htot.specs"
+! OUTPUT FILES CSFRAC_NAME.OUT AND CSNUM_NAME.OUT FOR FRACTIONAL ABUNDANCES AND NUMBER DENSITIES
+OUTPUT_FOLDER = "/lhome/silkem/CHEM/Output_nov2022/20211014_gridC_Mdot1e-6_v12-5_T_eps/models/model_2022-12-23h10-52-30/"
+SRC_CHEMISTRY = "src-IP-AP-HNR/"
