### 2024-10-21 875298d
Adding tutorial notebook on training a model
