### 2024-10-22 silkemaes/MACE#10 @trappitsch
Theses don't have to be automated, from the guidelines:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I saw that you have a `test.py`, which is a test case, however, I'm unclear if you'd consider this a test for the user to verify things work and what I would compare too. Would I compare to the annotated notebook? I think a bit of elaboration on this in the docs would be great such that the user can double check and run an example, compare it to what you have provided, and make sure it runs as intended. Hope this makes sense.
