### 2024-10-22 silkemaes/MACE#14 @silkemaes
Hi @richings,

You were correct to notice some files were missing or empty by accident. These issues should be fixed now.

I added a notebook version of the routine-script, which is similar to the notebook found on the documentation page. Because the documentation pages (and notebooks there) live on the main branch.

Hopefully now you are able to run the routine notebook without errors.
