### 2024-10-24 silkemaes/MACE#7 @richings
Hi @silkemaes

Great, thanks! That's fixed the link on the JOSS branch, although I think the link is still broken on the Readme on the main branch of the repository?
