### 2024-10-24 silkemaes/MACE#14 @richings
Hi @silkemaes

Thanks for the updates! I have tried running the routine.ipynb Jupyter Notebook that you have added on the JOSS branch. This has fixed the errors that I was finding earlier, and I can now run it further than I managed before. However, I am struggling to reproduce the same plots that you have in the notebook.

So I created a copy of the notebook from the repository and I restarted the kernel and cleared all outputs. I then tried to run through the notebook, but using the same pre-trained model that you provide, so I skip the "Set up the model" cell. I get to Stage 1, where it does the initial training with unscaled losses. It appears that it does train for 5 epochs, as you have in your copy of the notebook. However, the plot that it produces looks very different. When I do it, the plot only shows points at epoch 0.00, and the x-axis only extends from -0.05 to +0.05. But in your notebook, the plot that it produces after Stage 1 covers epochs of 0 to 13. I have attached a copy of the plot that I get when I run it below.

![routine_stage1_plot](https://github.com/user-attachments/assets/c8de2b67-424a-46de-8a14-bd673cac7ddc)

Am I doing something wrong here?

I then continued with Stage 2, where it continues the training with re-scaled losses. It trains for epochs 6 to 15, which is also the same as you have in your notebook. And the two plots that it produces here are closer to what you have, although the range on the x-axis is again a bit different. On my plots the first one covers epochs 0-5 and the second one 0-14 (which I think makes sense given how many epochs it has trained for), but on your notebook they cover epochs 0 to 30 and 0 to 40 respectively.

Btw in the Stage 2 plots, the losses suddenly jump (after epoch 4 in my plots, and after epochs 14-17 in your plots). Is this jump to be expected? Am I right in saying this is just because the losses are re-scaled after that epoch?

I was then able to save the model no problem. However, when I then tried the Testing the Model section, the cell under Performing the tests seems to take a long time. I gave it a few minutes but it didn't seem to make much progress. Is this expected? Although I noticed that when I run this cell it says "Selected Paths: 1000", but in your copy of the notebook it says "Selected Paths: 4" and "Selected test paths: 100", maybe this is why it is taking a lot longer when I run it?
