### 2024-10-25 86c5f51
Add extra data for tutorial
