### 2024-10-25 fe16e5b
Fixed issues with plotting in routine notebook and made sure it works correctly
