+++ pyproject.toml
@@ -28,7 +28,7 @@ dependencies = [
     "scipy",
     "torchode"
 ]
-requires-python = ">=3.7"
+requires-python = ">=3.8"
 
 [project.optional-dependencies]
 dev = ["black", "bumpver", "pip-tools", "pytest"]
