### 2024-10-25 ce6c5d5
making script routine.py compatible with routine.ipynb
