+++ input/example.in
@@ -24,16 +24,3 @@ elm = 0
 
 
 
-
-
-Name = 20241025_120806
-
-Name = 20241025_120806
-
-Name = 20241025_123012
-
-Name = 20241025_123835
-
-Name = 20241025_124525
-
-Name = 20241025_125032
