### 2024-10-25 silkemaes/MACE#14 @silkemaes
Hi @richings,

There were indeed still some issues with the routine notebook and the input files.

- Normally the plotting of the losses should work. There were general issues with plotting in python notebooks, that is always shows the plot, even without the ```plt.show()``` command. I normally made sure that now only the plots I aim to show are visible.
- The plotting should now work for stage 1 as well as stage 2.
- It is correct that the testing of the models didn't work properly, I found the hard coded number ```100``` somewhere, and removed this. I also added a bit more data: 13 samples in total, of which 10 for training + validating (with 70%-30% ratio) and 3 for testing. I of course know this is way too little data to properly train anything in machine learning, but here the data is just there to show the code works.

I hope with this you can continue.
