### 2024-10-25 291edf5
Add contribution guidlines to README
