+++ docs/src/index.rst
@@ -60,14 +60,17 @@ Developers & contributions
 *****
 
 **Developers**:
+
 * Silke Maes
 * Frederik De Ceuster
 
 **Scientific & technical advisors**:
+
 * Marie Van de Sande
 * Leen Decin
 
 **Contributors**:
+
 * Steven Rieder
 
 > Feel free to contribute to MACE via pull requests and discussions!
