### 2024-10-31 silkemaes/MACE#14 @richings
Hi @silkemaes

Great, thanks! That has fixed the errors I was finding, and I am now able to run the routine.ipynb notebook no problem and I can reproduce your results.

I then tried to run the load&test.ipynb notebook in the repository, this is the same as the one provided in the documentation here:

https://mace-code.readthedocs.io/en/latest/example/load%26test.html

I can run the "Loading the model" section no problem, and it produces the same plot that you had. However, when I then run the Testing the model section, it creates the first group of plots OK but part way through it gives an error:

ValueError: Unrecognized marker style 'none'

It looks like this is coming from matplotlib, if I look through the traceback I think it originates from line 75 of MACE/src/mace/CSE_0D/plotting.py where it has:

line, = ax1.plot(r,n_hat[:,idx], ls =ls, marker = marker, label = spec, ms = ms,  lw = lw)

It sounds like it might be trying to set the marker style as a string 'none' rather than as a None type python object?
