### 2024-10-31 silkemaes/MACE#16 @richings
JOSS Review: Additional comments on the manuscript
Hi @silkemaes

I have read through the JOSS Software Paper. This manuscript is well-written and summarises the functionality very well. I just had a few minor comments and suggestions:

- When you discuss the state of the field (e.g. in the Summary section), you say that MACE is "inspired on similar work in the literature (e.g. Holdship et al. 2021, Grassi et al. 2022, Sulzer & Buck 2023)". Do these other works also provide publicly-available packages, or is MACE the only available package that implements Machine Learning chemistry emulation like this? It would be good to clarify this in the paper.

- In Figure 2 you show an example of chemical abundance profiles from a trained MACE model that uses the int4 model from Maes et al (2024). Is this the same model that you used in the load&test.ipynb notebook example, e.g. from the documentation here:

https://mace-code.readthedocs.io/en/latest/example/load%26test.html

The plot in the notebook is almost the same as what you show in the paper, but there are some small differences, e.g. the curves for H2O and C10H2+ cross over in the notebook plot but not in Fig 2 of the paper. If this is the same model, are these differences to be expected?

- When you discuss the Example Case, you say that MACE gives an average speed-up of a factor of 26. It might be worth also giving some details of the machine that you ran these tests on to measure this speed up, for example does MACE give a bigger speed up if it is run on a GPU?
