### 2024-10-31 #7148 @richings
I had a few minor comments on the Software Paper. I have opened a new issue on the target repository [here](https://github.com/silkemaes/MACE/issues/16).
