### 2024-11-22 silkemaes/MACE#14 @silkemaes
Hi @richings,

I adjusted the marking type in the plotting script, though it is strange it throws an error since 'none' as a string should work as marker.

Anyway, it doesn't matter too much which marker is used there.

Hopefully you can proceed.

Thanks!
