### 2024-11-22 silkemaes/MACE#16 @silkemaes
Hi @richings,

Thank you for your feedback on the manuscript. I will implement it accordingly.

About the MACE model used in the figure of the manuscript and the example notebook, it is not entirely the same. They have he same hyperparameters though, but for the notebook, we trained it again to be compatible with the revised (and cleaned) code (the one submitted for JOSS and currently in branches 'main' and 'JOSS'). Hence, there are small differences and they are to be expected, since of course every trained ML model, though having the same hyperparameters, will be slightly different in weights.
