### 2024-11-28 silkemaes/MACE#14 @richings
Hi @silkemaes,

Great, many thanks! That has fixed the problem I was having, I am now able to run both of the annotated Jupyter Notebooks, and I can reproduce your results.

I will mark this issue as resolved.
