### 2025-01-02 silkemaes/MACE#10 @silkemaes
Hi @trappitsch, sorry I completely missed your reply on this issue.

First my best wishes to you for 2025!

The ```test.py``` script is not a testing script to verify whether things work or not, but it is part of the training. This is stated at the beginning of the script as follows:
```
This script contains functions to apply a trained MACE model on a test dataset,
i.e., test the trained model.

    - Test the model on 1 consecutive time step or on the evolution of the system.
    - Test the model on a full evolution of the system.
```

There are no tests (yet) to verify everything works correctly. As I understood correctly from the documentation online for JOSS submissions, tests are advised to include, but not obligatory.

Moreover, the code MACE provides is rather a method than a tool, and hence a test suite is less applicable here in our opinion.

For now, is it possible to proceed with the review without the tests?

If not, If you do require us to provide a test suite, we will try to do so.
