### 2025-01-02 silkemaes/MACE#10 @silkemaes
Moreover, the code MACE provides is rather a method than a tool, and hence a test suite is less applicable here in our opinion.

However, I you do require us to provide a test suite, we will try to do so.
