### 2025-01-02 908da52
Add info on documentation in README.md
