### 2025-01-02 silkemaes/MACE#16 @silkemaes
Hi @richings,

First my best wishes to you for 2025!

I will implement your feedback on the manuscript asap. In the mean time, I was wondering about your status/progress on reviewing the MACE code. Could you give me information on this?

Thank you!
