### 2025-01-07 silkemaes/MACE#16 @richings
Hi @silkemaes,

Happy New Year!

I have finished reviewing the rest of the MACE code, I am able to run the code no problem and I can reproduce the results in the examples that you show in the documentation, so that all looks good.

So it's just the manuscript that needs to be updated, once that's done I can then complete the review.
