### 2025-01-21 silkemaes/MACE#10 @trappitsch
Thanks @silkemaes for the clarification and sorry for the late reply. This looks reasonable to me, would be great if @JBorrow could have a look this to confirm.
