### 2025-01-22 #7148 @trappitsch
I am almost through. @JBorrow: Could you have a look at https://github.com/silkemaes/MACE/issues/10 and comment on the question for you? Thanks!
