### 2025-01-22 silkemaes/MACE#10 @JBorrow
That's correct, we do not require automated testing suites. However, we do require that there be some proof of accuracy - which I presume this code does by producing diagnostic plots?
