### 2025-01-22 silkemaes/MACE#10 @trappitsch
Correct, thanks @JBorrow. There's a whole section in the examples on testing MACE with plots and explanations.
