### 2025-01-30 #7148 @richings
Hi @JBorrow ! I've finished going through the code and that all looks good, I just had a few minor comments remaining on the Manuscript, see [this issue](https://github.com/silkemaes/MACE/issues/16) on the repository. Once the manuscript has been updated I can complete the review.
