### 2025-01-30 #7148 @trappitsch
@JBorrow Looking all good from my side, all issues (see references in my checklist) have been resolved. I don't consider the one that is still open a hindering issue for acceptance, see discussion in there. Furthermore, @silkemaes stated that she will make the change.
Thanks for the patience!
