### 2025-02-04 32740c5
Update paper.md according to comments of referees
