+++ paper.md
@@ -39,19 +39,21 @@ aas-journal: Astrophysical Journal
 ---
 
 # Summary
-Astrochemistry is the study of chemical species in astrophysical objects. By combining observations and theoretical predictions, the physical conditions of the object can be analysed as well as its chemical composition and yields, since chemistry is closely coupled to the dynamics of the environment. Most often, such astrophysical environments are known to be complex and out of thermodynamic equilibrium. For that reason, the chemical evolution is usually modelled using the chemical kinetics approach, in which a set of non-linear coupled ordinary different equations (ODEs) is solved for a given network of chemical species and reactions. For large chemical networks, however, this method is computationally slow. Therefore, we developed MACE, *a Machine-learning Approach to Chemistry Emulation*, inspired on similar work in the literature (e.g., @Holdship2021; @Grassi2022; @Sulzer2023). MACE is a PyTorch module that offers a trainable surrogate model that is able to emulate chemical kinetics in a shorter computation time than its classical analogue. More specifically, it provides a machine learning architecture, consisting of an autoencoder and a trainable ODE, implemented in PyTorch and [torchode](https://github.com/martenlienen/torchode) [@torchode]. Mathematically, MACE is give by 
+Astrochemistry is the study of chemical species in astrophysical objects. By combining observations and theoretical predictions, the physical conditions of the object can be analysed as well as its chemical composition and yields, since chemistry is closely coupled to the dynamics of the environment. Most often, such astrophysical environments are known to be complex and out of thermodynamic equilibrium. For that reason, the chemical evolution is usually modelled using a chemical kinetics approach, in which a set of non-linear coupled ordinary different equations (ODEs) is solved for a given network of chemical species and reactions. For large chemical networks, however, this method is computationally slow. Therefore, we developed MACE, *a Machine-learning Approach to Chemistry Emulation*, inspired by similar work in the literature, e.g., @Holdship2021; @Grassi2022; @Sulzer2023. Their code is publicly and can be found in their respective papers. However, not all of their code can be easily adapted or used as such. 
+
+MACE is a PyTorch module that offers a trainable surrogate model that is able to emulate chemical kinetics in a only 5% of the computation time of its classical analogue, in the current implementation. More speed up is expected when using GPUs instead of CPUs only. MACE provides a machine learning architecture, consisting of an autoencoder and a trainable ODE, implemented in PyTorch and [torchode](https://github.com/martenlienen/torchode) [@torchode]. Mathematically, MACE is give by 
 
 \begin{equation} \label{eq:mace}
 \hat{\textbf{\textit{n}}}(t) = \mathcal{D}\Big( G \big( \mathcal{E} (\textit{\textbf{n}}, \textit{\textbf{p}}),t \big) \Big), 
-\end{equation}
+\end{equation},
 
-where $\hat{\textbf{\textit{n}}}(t)$ represents the predicted chemical abundances at time $t$, more details can be found in @Maes2024 and in Fig. \ref{fig:MACE}, which shows a schematic representation of the architecture and flow of the emulator. MACE is developed in the context of circumstellar envelopes (CSEs) of evolved stars, but is flexible to be retrained on data of other astrophysical environments.
[…]
