### 2025-02-04 silkemaes/MACE#16 @silkemaes
Hi @richings,

I just adjusted the manuscript by implementing your comments as well as those of the other reviewer. I commit the changes to the JOSS branch.
