### 2025-02-06 silkemaes/MACE#16 @richings
Hi @silkemaes ,

Great, thanks! I've had a look through the updated manuscript, and this all looks good, you have addressed all of my comments very well!

I'll now mark this issue as resolved.
