### 2025-02-06 #7148 @richings
Hi @JBorrow ,

@silkemaes has updated the manuscript and has addressed all of the comments that I had, so all of the remaining issues on the repository have now been resolved.

I have also completed my JOSS check list, so everything looks good!

Let me know if there is anything else that you need?
