### 2025-03-11 #7148 @silkemaes
Hi @JBorrow ,

I double checked the affiliations of the others and added one extra. I merged the JOSS branch to the main branch to include all changes made thanks to the reviewers, and released this lastest version with the tag v1.1.0. Also this version is now available on Zenodo: DOI: 10.5281/zenodo.15007184.

If there is anything else I still need to take care of, please let me know.

Cheers,
Silke
