### 2025-03-18 #7148 @JBorrow
Hi @silkemaes, please could you make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper? They are different right now.
