### 2025-03-24 #7148 @silkemaes
Hi @JBorrow,

The co-authors on the JOSS paper are the same as the co-authors on the main publication of MACE, though not all authors have made a contribution to the MACE repository. However, the management, collaboration and funding they provided where necessary for MACE to be developed.

Could you maybe specify in which way the author list in the archive doesn't match the authors list of the JOSS paper?

Thanks!

Silke
