### 2025-03-24 #7148 @JBorrow
The title of 10.5281/zenodo.15007184 is 'silkemaes/MACE: Release after JOSS revision', it must be 'MACE: a Machine-learning Approach to Chemistry Emulation'
