### 2025-03-24 6e78184
Create CITATION.cff
