+++ CITATION.cff
@@ -1,0 +1,15 @@
+cff-version: 1.1.0
+message: "If you use this software, please cite it as below."
+authors:
+- family-names: Maes
+  given-names: Silke
+orcid: https://orcid.org/0000-0003-4159-9964
+- family-names: De Ceuster
+  given-names: Frederik
+- family-names: Van de Sande
+  given-names: Marie
+- family-names: Decin
+  given-names: Leen
+title:silkemaes/MACE: a Machine-learning Approach to Chemistry Emulation
+version: v1.1.0
+date-released: 2025-03-11
