### 2025-03-27 #7148 @silkemaes
Hi @JBorrow,

A few days ago I made a new release on GitHub with the correct title (the same as the manuscript). I see that on Zenodo the release title is not updated, though it redirects you to the correct release on GitHub. Is this okay to proceed? I don't see a way to remove the release completely from Zenodo, and updating the syncing between Zenodo and GitHub doesn't seem to help.

Best regards,
Silke
