### 2025-03-27 #7148 @JBorrow
Unfortunately we require that the title of the archive matches that of the paper. You should be able to simply edit the title of the archive using the yellow 'Edit' button in the top right:
