### 2025-03-27 #7148 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
