### 2025-03-28 #7148 @warrickball
Hi @silkemaes! I'm the AEiC (associate editor-in-chief) on this track so I'll take over and start checking that everything is ready for publication. I'll check a variety of things and probably open a pull request over the weekend with any editorial tweaks. So aside from checking the latest proof, you don't need to do anything until you hear from me again, which should be in the next few days.
