### 2025-03-28 #7148 @silkemaes
@warrickball, hi! And thank you.

I checked the proofs and found a small mistake from my side in lines 32-33. I now corrected it, pushed the changes to GitHub and created a new release. Zenodo automatically synced this release and created a new DOI (I hope this is not a problem): 10.5281/zenodo.15101491. The release version is still the same.

Cheers,
Silke
