### 2025-03-28 #7148 @warrickball
No problem. We don't require a new release for changes to the paper but it's easy enough to update the metadata anyway.
