### 2025-03-28 #7148 @warrickball
Hi @silkemaes, among my initial checks, I noticed that the author list of the archive (Maes, Rieder & De Ceuster) differs from the paper (Maes, De Ceuster, Van de Sande & Decin). Unless there's a particular reason for them to differ, they should be the same. I suspect the Zenodo author list has been generated automatically from the GitHub contributors, though that also means I'd appreciate your comments on why Steven Rieder isn't an author on the paper.

As Josh showed above, you can edit the metadata without creating a new archive.

While you're at it, you should add the ORCID information to the Zenodo authors.

Finally, in the paper itself it looks like you've added an AAS Journals DOI to the metadata. Was this specifically a joint submission with AAS Journals? It hasn't been mentioned before in the submission process so I presume not. (I can remove the extra metadata when I go through the article closely later.)
