### 2025-03-31 #7148 @silkemaes
Hi @warrickball, you were correct that the author list of on the archive was not correct. I now adjusted it accordingly and added the orcid IDs of the co-authors. Steven Rieder did not contribute to the development of MACE, only he once corrected a mistake in the `requirements.txt` via a pull request.
