### 2025-03-31 #7148 @warrickball
Thanks for these updates. I'm having a look at the paper now and hope to submit a PR with those fixes shortly.

> About the AAS Journals DOI, this refers to the scientific publication using MACE. So it is not as much a joint submission, but we want this publication to be associated to the JOSS publication, if that is possible.

You can (and should!) cite the AAS article but the association in the sidebar is a specific process where a paper is submitted to AAS Journals and JOSS simultaneously. You can find out more in [the announcement](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing). I'll double check that I've understood the AAS Journals side correctly but I'm pretty sure that this is *not* the case here.
