### 2025-03-31 52ca55e
Initial editorial tweaks for JOSS paper
