+++ paper.bib
@@ -1,7 +1,7 @@
 @ARTICLE{Holdship2021,
        author = {{Holdship}, J. and {Viti}, S. and {Haworth}, T.~J. and {Ilee}, J.~D.},
         title = "{Chemulator: Fast, accurate thermochemistry for dynamical models through emulation}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {astrochemistry, methods: numerical, methods: statistical, hydrodynamics, Physics - Computational Physics, Astrophysics - Astrophysics of Galaxies},
          year = 2021,
         month = sep,
@@ -19,7 +19,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Grassi2022,
        author = {{Grassi}, T. and {Nauman}, F. and {Ramsey}, J.~P. and {Bovino}, S. and {Picogna}, G. and {Ercolano}, B.},
         title = "{Reducing the complexity of chemical networks via interpretable autoencoders}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {astrochemistry, methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics, Physics - Chemical Physics},
          year = 2022,
         month = dec,
@@ -98,7 +98,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Gottlieb2022,
        author = {{Gottlieb}, C.~A. and {Decin}, L. and {Richards}, A.~M.~S. and {De Ceuster}, F. and {Homan}, W. and {Wallstr{\"o}m}, S.~H.~J. and {Danilovich}, T. and {Millar}, T.~J. and {Montarg{\`e}s}, M. and {Wong}, K.~T. and {McDonald}, I. and {Baudry}, A. and {Bolte}, J. and {Cannon}, E. and {De Beck}, E. and {de Koter}, A. and {El Mellah}, I. and {Etoka}, S. and {Gobrecht}, D. and {Gray}, M. and {Herpin}, F. and {Jeste}, M. and {Kervella}, P. and {Khouri}, T. and {Lagadec}, E. and {Maes}, S. and {Malfait}, J. and {Menten}, K.~M. and {M{\"u}ller}, H.~S.~P. and {Pimpanuwat}, B. and {Plane}, J.~M.~C. and {Sahai}, R. and {Van de Sande}, M. and {Waters}, L.~B.~F.~M. and {Yates}, J. and {Zijlstra}, A.},
         title = "{ATOMIUM: ALMA tracing the origins of molecules in dust forming oxygen rich M-type stars. Motivation, sample, calibration, and initial results}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {stars: AGB and post-AGB, stars: mass-loss, circumstellar matter, binaries: general, instrumentation: interferometers, astrochemistry, Astrophysics - Solar and Stellar Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2022,
         month = apr,
@@ -116,7 +116,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Maes2021,
        author = {{Maes}, S. and {Homan}, W. and {Malfait}, J. and {Siess}, L. and {Bolte}, J. and {De Ceuster}, F. and {Decin}, L.},
         title = "{SPH modelling of companion-perturbed AGB outflows including a new morphology classification scheme}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {stars: AGB and post-AGB, stars: winds, outflows, hydrodynamics, methods: numerical, Astrophysics - Solar and Stellar Astrophysics},
          year = 2021,
         month = sep,
@@ -134,7 +134,7 @@ archivePrefix = {arXiv},
[…]
