### 2025-03-31 #7148 @warrickball
I've just opened a [pull request](https://github.com/silkemaes/MACE/pull/18) with some initial editorial tweaks. If you're happy with them and merge them, I'll do another test run of our publication pipeline.

I'd also like to clarify this statement in the text:

> Using the MACE architecture, we are able to reproduce the 1D abundance profiles of 468 chemical species with an accuracy between 10 and 18, ...

I presume there's a unit missing from the accuracy? Perhaps it should be 10 and 18 *per cent*? Equation (2) suggests to me that it is indeed a fractional error (in log₁₀ **_n_**?) but I wasn't confident enough about that to propose it in the PR.
