### 2025-04-01 #7148 @arfon
> You can (and should!) cite the AAS article but the association in the sidebar is a specific process where a paper is submitted to AAS Journals and JOSS simultaneously. You can find out more in [the announcement](https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing). I'll double check that I've understood the AAS Journals side correctly but I'm pretty sure that this is not the case here.

Thanks Warrick. This is correct. @silkemaes – please simply cite your AAS article normally rather than using the special side-bar which is reserved for joint publications with AAS.
