### 2025-04-02 b5023cd
Improving explaination on error metric
