+++ paper.md
@@ -48,11 +48,7 @@ MACE is a PyTorch module that offers a trainable surrogate model that is able to
 \hat{\textbf{\textit{n}}}(t) = \mathcal{D}\Big( g \big( \mathcal{E} (\textit{\textbf{n}}, \textit{\textbf{p}}),t \big) \Big),
 \end{equation}
 
-<<<<<<< HEAD
-where $\hat{\textbf{\textit{n}}}(t)$ represents the predicted chemical abundances at time $t$. The symbols $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ represent the chemical abundances and physical parameters, respectively. The autoencoder consists of an encoder $\mathcal{E}$, able to map $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ to a latent space representation $\textbf{\textit{z}}$, and decoder $\mathcal{D}$, working the other way around. $G$ is a function that describes the evolution of the latent variables such that $\z(\Delta t) = G(\z, \Delta t)=\int_0^{\Delta t} g(\z){\rm d}t$, where $g$ represents the trainable ODE. A schematic representation of the architecture and flow of the emulator is given in Fig. \ref{fig:MACE} for a time step $\Delta t$. More details can also be found in @Maes2024. MACE is developed in the context of circumstellar envelopes (CSEs) of evolved stars, but is flexible to be retrained on data of other astrophysical environments.
-=======
-where $\hat{\textbf{\textit{n}}}(t)$ represents the predicted chemical abundances at time $t$. The symbols $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ represent the chemical abundances and physical parameters, respectively. The autoencoder consists of an encoder $\mathcal{E}$, able to map $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ to a latent space representation $\textbf{\textit{z}}$, and decoder $\mathcal{D}$, working the other way around. The function $g$ represents the trainable ODE, which needs to be solved for a time step $\Delta t$. A schematic representation of the architecture and flow of the emulator is given in \autoref{fig:MACE}. More details can also be found in @Maes2024. MACE is developed in the context of circumstellar envelopes (CSEs) of evolved stars, but is flexible to be retrained on data of other astrophysical environments.
->>>>>>> JOSS
+where $\hat{\textbf{\textit{n}}}(t)$ represents the predicted chemical abundances at time $t$. The symbols $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ represent the chemical abundances and physical parameters, respectively. The autoencoder consists of an encoder $\mathcal{E}$, able to map $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ to a latent space representation $\textbf{\textit{z}}$, and decoder $\mathcal{D}$, working the other way around. $G$ is a function that describes the evolution of the latent variables such that $\z(\Delta t) = G(\z, \Delta t)=\int_0^{\Delta t} g(\z){\rm d}t$, where $g$ represents the trainable ODE. A schematic representation of the architecture and flow of the emulator is given in Fig. \autoref{fig:MACE} for a time step $\Delta t$. More details can also be found in @Maes2024. MACE is developed in the context of circumstellar envelopes (CSEs) of evolved stars, but is flexible to be retrained on data of other astrophysical environments.
 
 ![Schematic representation of the architecture of MACE. The autoencoder is shown in blue, the time evolution happens in the red part of the emulator. [Adapted from @Maes2024.] \label{fig:MACE}](MACE.png)
 
@@ -64,13 +60,13 @@ Astrophysical objects or environments entail 3D dynamical processes, as well as 
 # Example case
[…]
