### 2025-04-02 #7148 @silkemaes
Hi @warrickball, I accepted your pull request on the JOSS branch and merged it with the main branch.

Further on the error metric, it is not a percentage be since it is a bit tricky with the logarithms, it is just a unitless number. I agree that it was written a bit confusingly. I now adapted this part in the manuscript to make it (hopefully) less confusing.

About the AAS publication, you can remove it then from the "In partnership with" sections of the paper (I don't think I can do it myself). Indeed a reference to the paper will suffice, also thanks to @arfon for clearing it out.
