### 2025-04-02 ca03716
Minor update README.md on example case
