### 2025-04-02 #7148 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
