### 2025-04-02 #7148 @warrickball
> Further on the error metric, it is not a percentage be since it is a bit tricky with the logarithms, it is just a unitless number. I agree that it was written a bit confusingly. I now adapted this part in the manuscript to make it (hopefully) less confusing.

Thanks for having a look. I'm still not sure the new explanation is clearer to me. Rather than try to work backwards from the equation, consider that the error shown in Figure 2 looks like it's of order 0.1 (away from the centre, anyway).

I'm also not sure the new explanation went onto the correct branch. Apparently the paper branch is `JOSS` but there's a change to text in commit https://github.com/silkemaes/MACE/commit/b5023cdd35156ebe0237575826317bdb4491b341 on the `main` branch. Should we be using that instead?
