### 2025-04-03 #7148 @silkemaes
Sorry for all the confusion! For the error shown in the bottom plot of figure 2, the same metric is used, but not summed over the different species, this is indeed not mentioned as such. Hence, if you would sum all errors in that plot, you should end up with a value around 10. I can adapt the text accordingly and make sure the changes are on the JOSS branch.

The diversion of branches is due to the release made.
