+++ paper.md
@@ -48,7 +48,7 @@ MACE is a PyTorch module that offers a trainable surrogate model that is able to
 \hat{\textbf{\textit{n}}}(t) = \mathcal{D}\Big( g \big( \mathcal{E} (\textit{\textbf{n}}, \textit{\textbf{p}}),t \big) \Big),
 \end{equation}
 
-where $\hat{\textbf{\textit{n}}}(t)$ represents the predicted chemical abundances at time $t$. The symbols $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ represent the chemical abundances and physical parameters, respectively. The autoencoder consists of an encoder $\mathcal{E}$, able to map $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ to a latent space representation $\textbf{\textit{z}}$, and decoder $\mathcal{D}$, working the other way around. The function $g$ represents the trainable ODE, which needs to be solved for a time step $\Delta t$. A schematic representation of the architecture and flow of the emulator is given in \autoref{fig:MACE}. More details can also be found in @Maes2024. MACE is developed in the context of circumstellar envelopes (CSEs) of evolved stars, but is flexible to be retrained on data of other astrophysical environments.
+where $\hat{\textbf{\textit{n}}}(t)$ represents the predicted chemical abundances at time $t$. The symbols $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ represent the chemical abundances and physical parameters, respectively. The autoencoder consists of an encoder $\mathcal{E}$, able to map $\textbf{\textit{n}}$ and $\textbf{\textit{p}}$ to a latent space representation $\textbf{\textit{z}}$, and decoder $\mathcal{D}$, working the other way around. $G$ is a function that describes the evolution of the latent variables such that $\z(\Delta t) = G(\z, \Delta t)=\int_0^{\Delta t} g(\z){\rm d}t$, where $g$ represents the trainable ODE. A schematic representation of the architecture and flow of the emulator is given in Fig. \autoref{fig:MACE} for a time step $\Delta t$. More details can also be found in @Maes2024. MACE is developed in the context of circumstellar envelopes (CSEs) of evolved stars, but is flexible to be retrained on data of other astrophysical environments.
 
 ![Schematic representation of the architecture of MACE. The autoencoder is shown in blue, the time evolution happens in the red part of the emulator. [Adapted from @Maes2024.] \label{fig:MACE}](MACE.png)
 
@@ -60,16 +60,16 @@ Astrophysical objects or environments entail 3D dynamical processes, as well as 
 # Example case
 In the context of circumstellar envelopes, we aim to build a comprehensive model, consisting of 3D hydrodynamics, chemistry, and radiation. This is needed since recently, complex structure (such as spiral, arcs, and bipolarity) are observed in CSEs around evolved, low- and intermediate mass stars, currently attributed to the presence of (sub-)stellar companions [@Nordhaus2006;@Decin2020;@Gottlieb2022]. In order to study these stars and their CSEs, taking into account companions, accurate 3-dimensional models are needed. Currently, only hydrodynamical modelling is done for these systems in 3D [e.g., @Maes2021; @Malfait2021], though the aim is to include chemistry as well as radiation [@Maes2022]. Starting with integrating chemistry and 3D hydrodynamics, this coupling will only be computationally feasible when a faster alternative to chemical kinetics exists.
 
-The development of MACE is a first step towards this coupling. Using the MACE architecture, we are able to reproduce the 1D abundance profiles of 468 chemical species with an accuracy between 10 and 18, averaged over 3000 tests, for MACE models with different sets of hyperparameters, for the following error metric:
[…]
