### 2025-04-03 #7148 @warrickball
A change to the text has introduced a LaTeX variable `\z`, which isn't valid. I'd guess you mean to just use `z` but I'm not sure if you wanted to style it like the variable `\textbf{\textit{z}}` already mentioned in the text. If you could fix that and comment with `@editorialbot generate pdf` to check that the paper builds, I'll do a quick proofread of the latest changes.

I'll try to wrap this up by the end of tomorrow, after which I'll be away for three weeks and expect to have almost zero capacity until I return. I think we're basically there though: the metadata deposit already looks valid, so I hope it's just this final tweak to the text.
