### 2025-04-04 2d56cc8
Fix latex mistake in manuscript
