### 2025-04-04 #7148 @warrickball
Congratulations @silkemaes, your paper has been published in JOSS! 🎉
